Add wormchain injective integration (#1260)
* Add wormchain injective integration

* Add metrics and logger for unknown wormchain

* Add wormchain rpc config

* Add evmos, kujira, osmosis to wormchainRpcProviders

* Add backfiller for tx-tracker with emitterChainId, emitterAddress and range time parameters


---------

+++ pipeline/cmd/backfiller/run.go
@@ -173,8 +173,8 @@ func publishVaa(ctx context.Context, push topic.PushFunc, queue chan *repository
 				Timestamp:        vaa.Timestamp,
 				UpdatedAt:        vaa.UpdatedAt,
 				TxHash:           vaa.TxHash,
-				Version:          vaa.Version,
-				Revision:         vaa.Revision,
+				Version:          uint16(vaa.Version),
+				Revision:         uint16(vaa.Revision),
 			}); err != nil {
 				logger.Error("Failed to push vaa", zap.Error(err))
 			} else {
